pyproject.toml
[tool.poetry]
name = "psxify"
version = "2025.3.29-2"
description = "A simple script to turn any image file into a PS1 texture."
authors = [ "Aaron Siegfried <siggnal@proton.me>" ]
license = "GPL"

[tool.poetry.dependencies]
python = "^3.9"
pillow = "5.4.1"

[tool.poetry.scripts]
psxify = 'psxify:main'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
